Make the rules dictate that the higher speed get's the hit
Before, I was subtracting the speed from the roll to hit...that became
clear to me that that made no sense as the faster you got the less
chance you were able to hit.  Instead, I added the speed to the roll.

I also added a comment explaining the method.