Biiiiiiig changes.  Read full description
First, changed the simple battle simulator to perform rolls based on
characters stats rather than a base value of 100.  This makes a huuuuuge
difference as Brian, the protagonist in the test code ALWAYS won.  Now
the formula for physical attacks is:

Dice roll between 1 and attacker's physical attack + attacker's speed
Dice roll between 1 and defender's physical defense + defender's speed

As long as the attacker's value is greater than defender's then the
attack hits.

Next up is the addition of weapons.  Weapons will now add to an
attackers physical damage.  Plans for including the weight in the result
of hit check will come.  Right now, a weapon can be of two types:
PHYSICAL or MAGICAL.  I have provided an enum to make reading the types
easier.

Added the ability to determine whether the attack is of a physical or
magical nature in the rules.  This was easy as you just have to check
the attacker's weapon's weapon type then call the appropriate method.

Added a new weapon, the Broadsword, to demonstrate the effectiveness of
equipping weapons.  Next will be adding armor.

Finally, (as far as I can remember) I broke the classes into appropriate
packages for organizations sake.